requirements.txt
appdirs==1.4.4
APScheduler==3.9.1
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bot-studio==1.4.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
django-restframework==0.0.1
djangorestframework==3.13.1
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
lxml==4.9.1
marshmallow==3.17.1
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytube==12.1.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
webdriver-manager==3.8.3
windows-curses==2.3.0
wsproto==1.2.0
youtube-comment-scraper-python==1.0.0
